server/chatbot/requirements.txt
flask==2.2.2
flask-cors==3.0.10
python-dotenv==0.19.2
langchain==0.0.250
langchain-google-genai==0.0.4
langchain-community==0.0.7
faiss-cpu==1.7.4
gunicorn==20.1.0
